refactor(tasks): extract ownership guards, assignment message and briefing protocols
Move four self-contained concerns out of the controller's tasks module into
sibling modules under agent-teams-controller/src/internal/, with no behaviour
change:

- taskOwnershipGuards.js — actor normalisation and the task ownership /
  deletion guards (normalizeActorName, isClearOwnerValue, assertKnownTaskActor,
  assertTaskOwnerMutation, assertTaskOwnerChange, assertTaskNotDeleted,
  readMutableTask, isSameMember, isSameTaskMember). resolveAgentTaskMutationActor
  and isLeadTaskActor become module-private; they had no callers outside the
  moved block.
- taskAssignmentMessage.js — buildAssignmentMessage, the assignment
  notification text.
- agentLanguage.js — buildMemberLanguageInstruction and its locale helpers.
  Named to mirror the existing src/shared/utils/agentLanguage.ts.
- briefingProtocols.js — the action-mode and process-registration protocol
  text plus the member formatting protocol. Not named "member*": the existing
  JSDoc says these are shared between lead (actionModeInstructions.ts,
  TeamProvisioningPromptBuilders.ts) and member briefings.

The code is moved verbatim, including the file's existing 4-space indentation
(agent-teams-controller/** is outside the repo-root prettier and eslint globs,
so a reindent would be unforced reformatting and would break byte-level
reviewability of the move). Review with
`git show <sha> --color-moved=dimmed-zebra`.

The public surface of tasks.js is unchanged: it still re-exports
buildActionModeProtocolText, MEMBER_DELEGATE_DESCRIPTION and
buildProcessProtocolText, so controller.js and both
agent-teams-controller.d.ts declaration files are untouched. The four requires
are inserted directly after the existing memberMessagingProtocol require.

buildMemberTaskProtocol is deliberately left in place even though its
neighbours move: it is under active change elsewhere and moving it would
manufacture a conflict for no size benefit.

284 lines moved out, 21 lines of requires back: tasks.js 1233 -> 970. No test
changes; the existing controller test suite (170 tests) passes unmodified,
which is the proof that the motion is behaviour-free.
